include testJs insidie Mock.js ffi script
Since purescript does not support local requires in ffi modules, we have
to include testJs in Mock.js:
https://github.com/purescript/purescript/issues/2343
